requirements.txt
allure-pytest==2.12.0
PyHamcrest==2.0.4
pytest==7.2.0
requests==2.28.1